+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - show "out of stock"
 - fix "quantity" showing up for variants
   - (label for "quantity to add to cart" which was hidden yesterday)
+- show for staff in product details the qty available for each variant
 
 
 ## 2018-10-01